proposal/pyproject.toml
[project]
name = "proposal"
version = "0.1.0"
description = ""
authors = [
    {name = "Sanchay Kumar",email = "dsanchaykumar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "langchain-groq (>=1.1.0,<2.0.0)",
    "langchain-classic (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "chromadb (>=1.3.5,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain (>=1.1.2,<2.0.0)",
    "langchain-tavily (>=0.2.15,<0.3.0)",
    "streamlit (>=1.52.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "proposal", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "coverage (>=7.13.0,<8.0.0)",
    "reportlab (>=4.4.5,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"